refactor: sql model
