+++ src/model/__init__.py
@@ -1,10 +1,8 @@
-from .major import Major
-from .pig import PIG, PIGStatus
-from .pig_member import PIGMember
-from .sig import SIG, SIGStatus
-from .sig_member import SIGMember
-from .user import User, UserRole, UserStatus
-from .sig_global_status import SIGGlobalStatus
-from .pig_global_status import PIGGlobalStatus
 from .article import Article, Board
 from .file_metadata import FileMetadata
+from .major import Major
+from .pig import PIG, PIGMember, PIGStatus
+from .pig_global_status import PIGGlobalStatus
+from .sig import SIG, SIGMember, SIGStatus
+from .sig_global_status import SIGGlobalStatus
+from .user import User, UserRole, UserStatus
