+++ src/model/pig.py
@@ -1,6 +1,5 @@
 from datetime import datetime, timezone
 from enum import Enum
-from typing import Optional
 
 from sqlmodel import CheckConstraint, Field, SQLModel, UniqueConstraint
 
@@ -15,17 +14,16 @@ class PIGStatus(str, Enum):
 class PIG(SQLModel, table=True):
     __tablename__ = "pig"  # type: ignore
     __table_args__ = (
-        UniqueConstraint("title", "year", "semester",
-                         name="uq_title_year_semester"),
+        UniqueConstraint("title", "year", "semester", name="uq_title_year_semester"),
         CheckConstraint("year >= 2025", name="ck_year_min"),
         CheckConstraint("semester IN (1, 2)", name="ck_semester_valid"),
     )
 
-    id: Optional[int] = Field(default=None, primary_key=True)
+    id: int = Field(default=None, primary_key=True)
 
     title: str = Field(nullable=False)
     description: str = Field(nullable=False)
-    content_id: Optional[int] = Field(foreign_key="article.id", unique=True)
+    content_id: int = Field(foreign_key="article.id", unique=True)
 
     status: PIGStatus = Field(nullable=False)
 
@@ -36,3 +34,18 @@ class PIG(SQLModel, table=True):
     updated_at: datetime = Field(default_factory=lambda: datetime.now(timezone.utc), nullable=False)
 
     owner: str = Field(foreign_key="user.id", nullable=False)
+
+
+class PIGMember(SQLModel, table=True):
+    __tablename__ = "pig_member"  # type: ignore
+    __table_args__ = (
+        UniqueConstraint("ig_id", "user_id", "status", name="uq_ig_user_status"),
+    )
+
+    id: int = Field(default=None, primary_key=True)
+
+    ig_id: int = Field(foreign_key="pig.id", nullable=False)
+    user_id: str = Field(foreign_key="user.id", nullable=False)
+    status: PIGStatus = Field(nullable=False)
+
+    created_at: datetime = Field(default_factory=lambda: datetime.now(timezone.utc), nullable=False)
